Little refactor
